requirements.txt
pydantic==2.9.2
fastapi==0.112.0
uvicorn[standard]==0.30.6
httpx==0.27.0
openai==1.45.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
google-generativeai==0.7.2
google-genai>=0.6.0
supabase==2.5.1
